+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3675,7 +3675,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KNS4KMN3JPPPP77JX560HY9R"),
+  id: z.string().default("01KNSDDEK65M3130AFHX1D10M5"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>