+++ packages/orq-rc/RELEASES.md
@@ -8159,3 +8159,13 @@ Based on:
 - [typescript v4.8.0-rc.0] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.0 - packages/orq-rc
+
+## 2026-04-09 15:26:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.1] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.1] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.1 - packages/orq-rc